create booking

+++ src/main/java/com/smartService/SmartServiceBookingAPIs/Repositories/ServiceRepository.java
@@ -3,4 +3,4 @@ package com.smartService.SmartServiceBookingAPIs.Repositories;
 import com.smartService.SmartServiceBookingAPIs.Entity.Services;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ServiceRepository extends JpaRepository<Services, Integer> {}
+public interface ServiceRepository extends JpaRepository<Services, Long> {}
